requirements.txt
flask 
redis 
celery 
face_recognition 
numpy 
Pillow
watchdog